# Experiment: Avoid `JSArrayConstr` for Varargs to optimize the Wasm backend