--- a/library/src/main/scala-new-collections/scala/scalajs/runtime/Compat.scala
+++ b/library/src/main/scala-new-collections/scala/scalajs/runtime/Compat.scala
@@ -12,6 +12,7 @@
 
 package scala.scalajs.runtime
 
+import scala.collection.immutable.ArraySeq
 import scala.collection.IterableOnce
 
 import scala.scalajs.js
@@ -20,6 +21,36 @@ private[runtime] object Compat {
 
   @inline def toScalaVarArgsImpl[A](array: js.Array[A]): Seq[A] =
     WrappedVarArgs.wrap(array)
+
+  @inline def toScalaVarArgsFromScalaArrayAnyRefImpl(array: Array[AnyRef]): Seq[AnyRef] =
+    new ArraySeq.ofRef(array)
+
+  @inline def toScalaVarArgsFromScalaArrayIntImpl(array: Array[Int]): Seq[Int] =
+    new ArraySeq.ofInt(array)
+
+  @inline def toScalaVarArgsFromScalaArrayDoubleImpl(array: Array[Double]): Seq[Double] =
+    new ArraySeq.ofDouble(array)
+
+  @inline def toScalaVarArgsFromScalaArrayLongImpl(array: Array[Long]): Seq[Long] =
+    new ArraySeq.ofLong(array)
+
+  @inline def toScalaVarArgsFromScalaArrayFloatImpl(array: Array[Float]): Seq[Float] =
+    new ArraySeq.ofFloat(array)
+
+  @inline def toScalaVarArgsFromScalaArrayCharImpl(array: Array[Char]): Seq[Char] =
+    new ArraySeq.ofChar(array)
+
+  @inline def toScalaVarArgsFromScalaArrayByteImpl(array: Array[Byte]): Seq[Byte] =
+    new ArraySeq.ofByte(array)
+
+  @inline def toScalaVarArgsFromScalaArrayShortImpl(array: Array[Short]): Seq[Short] =
+    new ArraySeq.ofShort(array)
+
+  @inline def toScalaVarArgsFromScalaArrayBooleanImpl(array: Array[Boolean]): Seq[Boolean] =
+    new ArraySeq.ofBoolean(array)
+
+  @inline def toScalaVarArgsFromScalaArrayUnitImpl(array: Array[Unit]): Seq[Unit] =
+    new ArraySeq.ofUnit(array)
 
   def toJSVarArgsImpl[A](seq: Seq[A]): js.Array[A] = {
     seq match {